[Refactor] @jaesunglee 리뷰 반영
