+++ core/common/android/src/main/kotlin/team/ppac/common/android/util/ClipBoard.kt
@@ -1,17 +1,19 @@
-package team.ppac.common.android.extension
+package team.ppac.common.android.util
 
 import android.content.ClipData
 import android.content.ClipboardManager
 import android.content.Context
 import android.graphics.Bitmap
-import android.widget.Toast
 import androidx.core.content.FileProvider
 import timber.log.Timber
 import java.io.File
 import java.io.FileOutputStream
 
 /* bitmap을 ClipBoard에 copy하는 함수, Menifest에 FileProvider 정의 필요*/
-fun Context.copyImageToClipBoard(bitmap: Bitmap): Boolean {
+fun Context.copyImageToClipBoard(
+    bitmap: Bitmap,
+    onFailure: (throwable: Throwable) -> Unit = { },
+): Boolean {
     return runCatching {
         // 비트맵 File로 저장
         val imagesFolder = File(cacheDir, "images").apply {
@@ -32,12 +34,11 @@ fun Context.copyImageToClipBoard(bitmap: Bitmap): Boolean {
         )
 
         // 클립보드에 이미지 Uri 복사
-        val clipboard: ClipboardManager =
-            getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+        val clipboard = getSystemService(Context.CLIPBOARD_SERVICE) as? ClipboardManager ?: throw ClassCastException()
         val clip = ClipData.newUri(contentResolver, "Image", uri)
         clipboard.setPrimaryClip(clip)
     }.onFailure {
-        Toast.makeText(this, "기기의 저장공간을 확인해주세요!", Toast.LENGTH_LONG).show()
+        onFailure(it)
         Timber.e(it)
     }.isSuccess
 }
